+++ tagging.tf
@@ -1,23 +1,30 @@
 locals {
-  tags = "${join(" ", formatlist("%s=%s", keys(var.tags), values(var.tags)))}"
+  tags = join(" ", formatlist("%s=%s", keys(var.tags), values(var.tags)))
 }
 
+data "azurerm_client_config" "current" {}
+
 resource "null_resource" "tags" {
+
+  count = var.nb_resources
+
   # Code found here https://docs.microsoft.com/en-us/azure/azure-resource-manager/resource-group-using-tags#code-try-22
   provisioner "local-exec" {
+    interpreter = ["/bin/bash", "-c"]
     command = <<CMD
-      current_tags=""
-      if "${var.behavior == "merge" ? "true" : "false"}" == "true"; then
-        current_tags=$(az resource show --ids ${var.resource_id} -o json --query "@.tags" | tr -d '"{},' | sed 's/: /=/g');
-      fi
-      az resource tag --tags $current_tags ${local.tags} --ids ${var.resource_id}
+      ${templatefile("${path.module}/script/tag.sh.tmpl", { behavior = var.behavior,
+    resource_id     = var.resource_ids[count.index],
+    tags            = local.tags,
+    subscription_id = data.azurerm_client_config.current.subscription_id
+  }
+)}
     CMD
-  }
+}
 
-  triggers {
-    "resource"    = "${var.resource_id}"
-    "tags"        = "${jsonencode(var.tags)}"
-    "force-apply" = "${var.force ? uuid() : "false"}"
-    "behavior"    = "${var.behavior}"
-  }
+triggers = {
+  resource    = var.resource_ids[count.index]
+  tags        = jsonencode(var.tags)
+  force-apply = var.force ? uuid() : false
+  behavior    = var.behavior
+}
 }