+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # v2.0.0 - Unreleased
 
-
+Changed:
+  * AZ-157: New input nb_resources to count the number of resources. Change resource_id to resource_ids.
+  
 Breaking
   * AZ-94: Upgrade to terraform 0.12