+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# v2.0.0 - Unreleased
+
+Changed:
+  * AZ-157: New input nb_resources to count the number of resources. Change resource_id to resource_ids.
+  
+Breaking
+  * AZ-94: Upgrade to terraform 0.12
+  
 # v1.0.0 - 2019-05-06
 
 Added